py-fastapi/requirements.txt
fastapi==0.104.0
Jinja2==3.1.2
PyJWT==2.6.0
uvicorn[standard]==0.23.2
